Adding handler of signal for each class

+++ src/lib/syscalls/SysCalls.h
@@ -25,10 +25,7 @@ inline pid_t fork() {
 
 	pid_t id = ::fork();
 
-	if (id < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (id < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a fork()");
 	}
 
@@ -40,10 +37,7 @@ inline pid_t getpid() {
 
 	pid_t result = ::getpid();
 
-	if (result < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (result < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a getpid()");
 	}
 
@@ -62,20 +56,14 @@ inline void execv(const string &execPath, vector<string> params) {
 	auto status = ::execv(execPath.c_str(), args);
 	perror("execError");
 
-	if (status < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (status < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a execv()");
 	}
 }
 
 inline int mknod(const char *pathname, mode_t mode, dev_t dev) {
 	int response = ::mknod(pathname, mode, dev);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a mknod()");
 	}
 	return response;
@@ -83,10 +71,7 @@ inline int mknod(const char *pathname, mode_t mode, dev_t dev) {
 
 inline int unlink(const char *pathname) {
 	int response = ::unlink(pathname);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a unlink()");
 	}
 	return response;
@@ -94,10 +79,7 @@ inline int unlink(const char *pathname) {
 
 inline int close(int fd) {
 	int response = ::close(fd);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a close()");
 	}
 	return response;
@@ -105,10 +87,7 @@ inline int close(int fd) {
 
 inline int open(const char *pathname, int flags, int otherFlags) {
 	int response = ::open(pathname, flags, otherFlags);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a open()");
 	}
 	return response;
@@ -116,10 +95,7 @@ inline int open(const char *pathname, int flags, int otherFlags) {
 
 inline int open(const char *pathname, int flags) {
 	int response = ::open(pathname, flags);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a open()");
 	}
 	return response;
@@ -127,10 +103,7 @@ inline int open(const char *pathname, int flags) {
 
 inline ssize_t read(int fd, void *buf, size_t count) {
 	int response = ::read(fd, buf, count);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a read()");
 	}
 	return response;
@@ -138,10 +111,7 @@ inline ssize_t read(int fd, void *buf, size_t count) {
 
 inline ssize_t write(int fd, const void *buf, size_t count) {
 	int response = ::write(fd, buf, count);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a write()");
 	}
 	return response;
@@ -149,10 +119,7 @@ inline ssize_t write(int fd, const void *buf, size_t count) {
 
 inline off_t lseek(int fd, off_t offset, int whence) {
 	off_t response = ::lseek(fd, offset, whence);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a lseek()");
 	}
 	return response;
@@ -160,10 +127,7 @@ inline off_t lseek(int fd, off_t offset, int whence) {
 
 inline key_t ftok(const char *pathname, int proj_id) {
 	key_t response = ::ftok(pathname, proj_id);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a ftok()");
 	}
 	return response;
@@ -171,10 +135,7 @@ inline key_t ftok(const char *pathname, int proj_id) {
 
 inline int semget(key_t key, int nsems, int semflg) {
 	int response = ::semget(key, nsems, semflg);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a semget()");
 	}
 	return response;
@@ -182,10 +143,7 @@ inline int semget(key_t key, int nsems, int semflg) {
 
 inline int semop(int semid, struct sembuf *sops, size_t nsops) {
 	int response = ::semop(semid, sops, nsops);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a semop()");
 	}
 	return response;
@@ -193,10 +151,7 @@ inline int semop(int semid, struct sembuf *sops, size_t nsops) {
 
 inline int shmget(key_t key, size_t size, int shmflg) {
 	int response = ::shmget(key, size, shmflg);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a shmget()");
 	}
 	return response;
@@ -204,10 +159,7 @@ inline int shmget(key_t key, size_t size, int shmflg) {
 
 inline void *shmat(int shmid, const void *shmaddr, int shmflg) {
 	void* response = ::shmat(shmid, shmaddr, shmflg);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a shmat()");
 	}
 	return response;
@@ -215,10 +167,7 @@ inline void *shmat(int shmid, const void *shmaddr, int shmflg) {
 
 inline int semctl(int semid, int senum, int cmd) {
 	int response = ::semctl(semid, senum, cmd);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a semctl()");
 	}
 	return response;
@@ -226,10 +175,7 @@ inline int semctl(int semid, int senum, int cmd) {
 
 inline int shmdt(const void *shmaddr) {
 	int response = ::shmdt(shmaddr);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a shmdt()");
 	}
 	return response;
@@ -237,10 +183,7 @@ inline int shmdt(const void *shmaddr) {
 
 inline int shmctl(int shmid, int cmd, struct shmid_ds *buf) {
 	int response = ::shmctl(shmid, cmd, buf);
-	if (response < 0) {
-		if (errno == EINTR){
-			throw SigIntException();
-		}
+	if (response < 0 && errno != EINTR) {
 		throw SysCallException("Error trying to execute a shmctl()");
 	}
 	return response;
